refactor: move public params and commitment
